backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
langchain-qdrant==0.2.0
langchain-community==0.3.0
qdrant-client==1.12.0
pypdf==5.0.0
requests==2.32.3
pydantic==2.9.0
